new component in a new file
